feat(money): use MinorUnit as storage type
Replace raw Int64 storage with the MinorUnit wrapper, which rejects
Int64.min at construction. Eliminates scattered Int64.min guards —
the invariant is now enforced once at the parsing boundary.

+++ Sources/SwiftMoney/AnyMoney/AnyMoney+Decimal.swift
@@ -10,7 +10,7 @@ extension AnyMoney {
     /// ```
     @inlinable
     public var decimalValue: Decimal {
-        Decimal(minorUnits) / Decimal(minimalQuantisation.int64Value)
+        Decimal(Int64(minorUnits)) / Decimal(minimalQuantisation.int64Value)
     }
 }
 #endif
